# Conflicts:
#	nodejs/src/cdp/cdp-services.ts
#	nodejs/src/cdp/services/hogflows/hogflow-executor.service.ts

+++ products/data_warehouse/backend/models/datawarehouse_saved_query.py
@@ -9,7 +9,6 @@ from django.core.exceptions import ValidationError
 from django.db import models, transaction
 
 import structlog
-from dlt.common.normalizers.naming.snake_case import NamingConvention
 
 from posthog.schema import DataWarehouseSavedQueryOrigin, HogQLQueryModifiers
 
@@ -22,6 +21,7 @@ from posthog.hogql.database.s3_table import DataWarehouseTable as HogQLDataWareh
 from posthog.exceptions_capture import capture_exception
 from posthog.models.utils import CreatedMetaFields, DeletedMetaFields, UpdatedMetaFields, UUIDTModel
 from posthog.sync import database_sync_to_async
+from posthog.temporal.data_imports.naming_convention import NamingConvention
 
 from products.data_warehouse.backend.models.util import (
     CLICKHOUSE_HOGQL_MAPPING,
@@ -83,7 +83,7 @@ class DataWarehouseSavedQuery(CreatedMetaFields, UUIDTModel, UpdatedMetaFields, 
     external_tables = models.JSONField(default=list, null=True, blank=True, help_text="List of all external tables")
     query = models.JSONField(default=dict, null=True, blank=True, help_text="HogQL query")
     status = models.CharField(
-        null=True, choices=Status.choices, max_length=64, help_text="The status of when this SavedQuery last ran."
+        null=True, choices=Status, max_length=64, help_text="The status of when this SavedQuery last ran."
     )
     last_run_at = models.DateTimeField(
         null=True,
@@ -116,7 +116,7 @@ class DataWarehouseSavedQuery(CreatedMetaFields, UUIDTModel, UpdatedMetaFields, 
     )
 
     origin = models.CharField(
-        choices=Origin.choices, help_text="Where this SavedQuery is created.", default=None, null=True, blank=True
+        choices=Origin, help_text="Where this SavedQuery is created.", default=None, null=True, blank=True
     )
 
     is_test = models.BooleanField(
@@ -195,23 +195,26 @@ class DataWarehouseSavedQuery(CreatedMetaFields, UUIDTModel, UpdatedMetaFields, 
         from products.data_warehouse.backend.data_load.saved_query_service import delete_saved_query_schedule
         from products.data_warehouse.backend.models.modeling import DataWarehouseModelPath
 
-        with transaction.atomic():
-            self.sync_frequency_interval = None
-            self.last_run_at = None
-            self.latest_error = None
-            self.status = None
-            self.is_materialized = False
+        self.sync_frequency_interval = None
+        self.last_run_at = None
+        self.latest_error = None
+        self.status = None
+        self.is_materialized = False
 
-            # delete the materialized table reference
-            if self.table is not None:
-                self.table.soft_delete()
-                self.table_id = None
+        should_delete_saved_query_schedule: bool = False
+        try:
+            with transaction.atomic():
+                # delete the materialized table reference
+                if self.table is not None:
+                    self.table.soft_delete()
+                    self.table_id = None
 
-            delete_saved_query_schedule(self)
-
-            self.save()
-
-            DataWarehouseModelPath.objects.filter(team=self.team, path__lquery=f"*{{1,}}.{self.id.hex}").delete()
+                should_delete_saved_query_schedule = True
+                self.save()
+                DataWarehouseModelPath.objects.filter(team=self.team, path__lquery=f"*{{1,}}.{self.id.hex}").delete()
+        finally:
+            if should_delete_saved_query_schedule:
+                delete_saved_query_schedule(self)
 
     def soft_delete(self):
         self.deleted = True
@@ -265,6 +268,9 @@ class DataWarehouseSavedQuery(CreatedMetaFields, UUIDTModel, UpdatedMetaFields, 
 
     @property
     def s3_tables(self):
+        return self.get_s3_tables()
+
+    def get_s3_tables(self, database=None):
         from posthog.hogql.context import HogQLContext
         from posthog.hogql.database.database import Database
         from posthog.hogql.parser import parse_select
@@ -277,8 +283,7 @@ class DataWarehouseSavedQuery(CreatedMetaFields, UUIDTModel, UpdatedMetaFields, 
             team_id=self.team.pk,
             enable_select_queries=True,
             modifiers=create_default_modifiers_for_team(self.team),
-            # KLUDGE: Should accept this as a parameter to avoid rebuilding it everytime this is called
-            database=Database.create_for(self.team.pk),
+            database=database or Database.create_for(self.team.pk),
         )
 
         query = self.query or {}
@@ -299,7 +304,7 @@ class DataWarehouseSavedQuery(CreatedMetaFields, UUIDTModel, UpdatedMetaFields, 
 
     @property
     def normalized_name(self):
-        return NamingConvention().normalize_identifier(self.name)
+        return NamingConvention.normalize_identifier(self.name)
 
     @property
     def url_pattern(self):
